Don't cast gcim output to CimInstance[] in get-*
On macOS (and presumably Linux), invoke-command {get-ciminstance}
returns PSObject or Object[] of PSObject and casting to CimInstance
errors, so the cast means these functions cannot be used outside of
Windows.

The minimum change to stop the cast to CimInstance would be to cast to
Object[], but Array makes the goal of ensuring an array rather than a
single object is returned clearer.

For comparison, Windows returns CimInstance or Object[] of CimInstance
rather than CimInstance[], so this change means the cast will become a
no-op in most situations and only change the array's type in others.